Final resbumission after Kovan and Rinkeby Orcalize testing
