+++ src/components/resolvers/Rinkeby/index.js
@@ -1,1 +1,1 @@
-export default ["0x16fD6e2E1C4afB9C4e7B901141706596317e4ceB","0xD9D40Ca7BedcC47E5D4084c4011dAc50566F8d4f"]
+export default ["0x16fD6e2E1C4afB9C4e7B901141706596317e4ceB","0x2930Cf9EE8E03C3E06Fa1828cCD8E371323Fde0f"]
